docs(adr): note workspaceType will be deprecated in future
The workspaceType configuration option is currently in use and works
correctly (creating Cargo.toml scaffolding for faster rust-analyzer
initialization). However, it will be deprecated in a future release
per the "Minimal File Creation" principle.

Users should be aware that migration to initializationOptions.linkedProjects
will be required when the deprecation is implemented.